solidity implementation for ics23
Code is splitted in 4 different libraries, following the layout of
functions and files of the go implementation
- Ops -> go/ops.go
- Proof -> go/proof.go
- Compress -> go/compress.go
- Ics23 -> go/ics23.go

Missing parts of the current implementation:
- sha512 hash option: solidity does not have built-in function for this
hash. A solidity implementation would utterly inefficient and very
costly (gas-wise)
- compress functions: the implementation will very likely be used to
verification, not for proof generation
- ProofSpec for tendermint and iavl definition: again, this
implementation is meant to be used to verify, not to provide proofs.

The implementation also provides a framework to run the same unit-test
cases that the go implementation has.
go-ethereum package provides a ``SimulatedBackend`` that can be used to
deploy a contract and call functions. UnitTest-specific contracts are
included in the libraries sol files.
The current go-ethereum version is the last with a working ``abigen``
command.

This commit also refactors the go project to make room for two go
modules in the same repository