Check for placeholder nodes in proof verification
Certain proofs (like those for a sparse Merkle tree) can contain empty nodes which are equal to the
EmptyChild, but currently these will cause verification to fail for non-existence proofs if they
appear where a real sibling node would be invalid, i.e. on a path that must be the leftmost or
rightmost path of a subtree.

This adds a check to detect and ignore such nodes during verification of left- and rightmost paths,
as well as unit tests.