fix empty branch check
The code erroneously assumed that the inner op has the correct padding for the side it is checking.
e.g. during IsLeftMost:
when hasPadding(step) == false, leftBranchesAreEmpty(step, 0) is called, but 0 indicates a
left-branch, and hasPadding == false implies this is _not_ a left branch.
Instead, we must detect the branch index from the op's padding, since the caller does not know
which, if any, branch this padding is valid for.