feat: enable no_std support for `pbjson` (#158)
Closes: #156

Addresses the `no_std` compatibility issue with `serde` feature in `ibc-rs`. This is caused by the recent implementation of ProtoJSON serialization and deserialization [0] for the `ics23` Protobuf definitions using `pbjson`, and then re-exporting the ics23 type [1] in `ibc-proto-rs`. Some of our users by then (starting from IBC-rs v0.41.0) are experiencing compilation errors. [2]

To meet this immediate need [3] and the lack of activity in the `pbjson` crate for months, we have taken the initiative to feature `no_std` support in the `informalsystems-pbjson` crate and have it published.

[0] https://github.com/cosmos/ics23/pull/146
[1] https://github.com/cosmos/ibc-proto-rs/pull/92
[2] https://github.com/cosmos/ibc-rs/issues/741
[3] https://github.com/cosmos/ibc-proto-rs/issues/98#issuecomment-1653311451

---

* feat: enable no_std support for pbjson

* fix: get serde feature work with no-std

* deps: use informalsystems-pbjson v0.6.0

* deps: use informalsystems-pbjson v0.6.0